Integrate splitted widgets into the AppMain
