+++ lib/helpers/pages_fixtures.dart
@@ -1,0 +1,36 @@
+import '../models/page.dart';
+
+class CourseList {
+  static List<Page> lists = [
+    Page(
+      title: "Data Science",
+      description:
+          "Launch your career in data science. A sweet-cource introduction to data science, develop and taught by leading professors.",
+    ),
+    Page(
+      title: "Machine Learning",
+      description:
+          "This specialization from leading researchers at university of washington introduce to you to the exciting high-demand field of machine learning ",
+    ),
+    Page(
+      title: "Big Data",
+      description:
+          "Drive better bussiness decision with an overview OF how big data is organised  and intepreted. Apply insight to real-world problems and question",
+    ),
+    Page(
+      title: "Data Science",
+      description:
+          "Launch your career in data science. A sweet-cource introduction to data science, develop and taught by leading professors.",
+    ),
+    Page(
+      title: "Machine Learning",
+      description:
+          "This specialization from leading researchers at university of washington introduce to you to the exciting high-demand field of machine learning ",
+    ),
+    Page(
+      title: "Big Data",
+      description:
+          "Drive better bussiness decision with an overview OF how big data is organised  and intepreted. Apply insight to real-world problems and question",
+    ),
+  ];
+}
