--- lib/helpers/quad_clipper.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-
-class QuadClipper extends CustomClipper<Rect> {
-  @override
-  Rect getClip(Size size) {
-    Rect rect = Rect.fromLTRB(0.0, 0.0, size.width / 2, size.height / 2);
-    return rect;
-  }
-
-  @override
-  bool shouldReclip(CustomClipper<Rect> oldClipper) {
-    return true;
-  }
-}
