+++ lib/main.dart
@@ -1,4 +1,17 @@
-import 'package:chiudiamo_il_cerchio/screens/home.dart';
 import 'package:flutter/material.dart';
 
-void main() => runApp(HomeScreen());
+import './screens/home.dart';
+
+void main() => runApp(AppMain());
+
+class AppMain extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Chiudiamo il cerchio!',
+      home: Scaffold(
+        body: HomeScreen(),
+      ),
+    );
+  }
+}
